+++ src/main/java/com/example/knu_connect/domain/user/entity/User.java
@@ -1,5 +1,6 @@
 package com.example.knu_connect.domain.user.entity;
 
+import com.example.knu_connect.domain.basetime.BaseTimeEntity;
 import com.example.knu_connect.domain.user.entity.enums.*;
 import jakarta.persistence.*;
 import lombok.*;
